Lots more WIP stuff on imputation and fixing var types in modules. Also added a testing notebook for comparing final_US, complete_US, and imputed stock data to influence decisions

+++ minos/modules/nutrition.py
@@ -88,7 +88,7 @@ class Nutrition(Base):
         newWaveNutrition = pd.DataFrame(newWaveNutrition, columns=["nutrition_quality"])
         # Set index type to int (instead of object as previous)
         newWaveNutrition.index = (newWaveNutrition.index.astype(int))
-        #newWaveNutrition['nutrition_quality'] = newWaveNutrition['nutrition_quality'].astype(float)
+        newWaveNutrition['nutrition_quality'] = newWaveNutrition['nutrition_quality'].astype(int)
 
         # Draw individuals next states randomly from this distribution.
         # Update population with new income
